feat: parallel
