+++ pkg/scraper/scraper.go
@@ -2,7 +2,6 @@ package scraper
 
 import (
 	"context"
-	"html"
 	"log"
 	"time"
 
@@ -77,29 +76,12 @@ func (s *Scraper) getVideoList(ids []string) (*youtube.VideoListResponse, error)
 	return res, nil
 }
 
-type Channel struct {
-	ID    string `json:"id"`
-	Title string `json:"title"`
-	URL   string `json:"url"`
-}
-
-type Video struct {
-	Channel     *Channel `json:"channel"`
-	ID          string   `json:"id"`
-	ObjectID    string   `json:"objectID"`
-	PublishedAt int64    `json:"publishedAt"`
-	Title       string   `json:"title"`
-	URL         string   `json:"url"`
-}
-
 type ScrapeOptions struct {
 	All             bool
 	PublishedBefore time.Time
 }
 
 func (s *Scraper) Scrape(channelID string, opts *ScrapeOptions) []*Video {
-	log.Printf("Scrape https://www.youtube.com/channel/%s", channelID)
-
 	var (
 		date      = opts.PublishedBefore
 		pageToken = ""
@@ -107,8 +89,8 @@ func (s *Scraper) Scrape(channelID string, opts *ScrapeOptions) []*Video {
 	)
 
 	for {
-		log.Printf(`published_before: "%s", page_token: "%s"`,
-			date.Format(time.RFC3339), pageToken)
+		log.Printf(`channel_id: "%s", published_before: "%s", page_token: "%s"`,
+			channelID, date.Format(time.RFC3339), pageToken)
 
 		searchOpts := &searchOptions{
 			pageToken:       pageToken,
@@ -133,37 +115,7 @@ func (s *Scraper) Scrape(channelID string, opts *ScrapeOptions) []*Video {
 		}
 
 		for _, item := range res.Items {
-			rawPublishedAt := item.Snippet.PublishedAt
-
-			if item.LiveStreamingDetails != nil {
-				if item.LiveStreamingDetails.ActualStartTime != "" {
-					rawPublishedAt = item.LiveStreamingDetails.ActualStartTime
-				} else if item.LiveStreamingDetails.ScheduledStartTime != "" {
-					rawPublishedAt = item.LiveStreamingDetails.ScheduledStartTime
-				}
-			}
-
-			publishedAt, err := time.Parse(time.RFC3339, rawPublishedAt)
-			if err != nil {
-				publishedAt = time.Now()
-			}
-
-			channel := &Channel{
-				ID:    item.Snippet.ChannelId,
-				Title: item.Snippet.ChannelTitle,
-				URL:   "https://www.youtube.com/channel/" + item.Snippet.ChannelId,
-			}
-
-			video := &Video{
-				Channel:     channel,
-				ID:          item.Id,
-				ObjectID:    item.Id,
-				PublishedAt: publishedAt.Unix(),
-				Title:       html.UnescapeString(item.Snippet.Title),
-				URL:         "https://www.youtube.com/watch?v=" + item.Id,
-			}
-
-			results = append(results, video)
+			results = append(results, normalize(item))
 		}
 
 		if !opts.All || (pageToken == "" && len(searchRes.Items) < 1) {
